Saving Instance State.
The instance of the daily meal fragment is saved so it can be restored
after the application has been "killed" by the android system without
crashing the whole app.

OPEN - task #3: Crash when trying to resume application when a meal plan
is opened.  http://github.com/atomfrede/mensa-ubi/issues/issue/3